+++ candidates/urls.py
@@ -1,8 +1,6 @@
 from django.urls import path
 from candidates.views import (
     index,
-    JobsListView,
-    JobsDetailView,
     about,
     profile,
     toggle_save_vacancy,
@@ -13,8 +11,6 @@ from candidates.views import (
 
 urlpatterns = [
     path("", index, name="index"),
-    path("jobs/", JobsListView.as_view(), name="jobs_list"),
-    path("jobs/<int:pk>/", JobsDetailView.as_view(), name="jobs_detail"),
     path("about/", about, name="about"),
     path("profile/", profile, name="profile"),
     path("toggle-save-vacancy/<int:pk>/", toggle_save_vacancy, name="toggle_save_vacancy"),
