+++ candidates_fixture.json
@@ -1,51 +1,75 @@
 [
   {
-    "model": "candidates.company",
+    "model": "candidates.candidate",
+    "pk": 1,
+    "fields": {
+      "password": "pbkdf2_sha256$260000$something$hashedvalue",
+      "last_login": null,
+      "is_superuser": true,
+      "username": "admin",
+      "first_name": "Admin",
+      "last_name": "Root",
+      "email": "admin@example.com",
+      "is_staff": true,
+      "is_active": true,
+      "date_joined": "2025-04-01T00:00:00",
+      "current_role": "E",
+      "groups": [],
+      "user_permissions": []
+    }
+  },
+  {
+    "model": "employers.company",
     "pk": 1,
     "fields": {
       "name": "Zeno Systems",
       "description": "Pioneering decentralized networks.",
-      "website": "https://zeno.io"
+      "website": "https://zeno.io",
+      "owner": 1
     }
   },
   {
-    "model": "candidates.company",
+    "model": "employers.company",
     "pk": 2,
     "fields": {
       "name": "CloudForge Labs",
       "description": "Next-gen cloud infrastructure services.",
-      "website": "https://cloudforge.dev"
+      "website": "https://cloudforge.dev",
+      "owner": 1
     }
   },
   {
-    "model": "candidates.company",
+    "model": "employers.company",
     "pk": 3,
     "fields": {
       "name": "LuminaSoft",
       "description": "Human-centered productivity tools.",
-      "website": "https://luminasoft.com"
+      "website": "https://luminasoft.com",
+      "owner": 1
     }
   },
   {
-    "model": "candidates.company",
+    "model": "employers.company",
     "pk": 4,
     "fields": {
       "name": "HyperNetics",
       "description": "Real-time analytics for enterprise.",
-      "website": "https://hypernetics.ai"
+      "website": "https://hypernetics.ai",
+      "owner": 1
     }
   },
   {
-    "model": "candidates.company",
+    "model": "employers.company",
     "pk": 5,
     "fields": {
       "name": "NextBlock",
       "description": "Blockchain infrastructure solutions.",
-      "website": "https://nextblock.io"
+      "website": "https://nextblock.io",
+      "owner": 1
     }
   },
   {
-    "model": "candidates.vacancy",
+    "model": "employers.vacancy",
     "pk": 1,
     "fields": {
       "created_at": "2025-04-01T09:45:00",
@@ -58,7 +82,7 @@
     }
   },
   {
-    "model": "candidates.vacancy",
+    "model": "employers.vacancy",
     "pk": 2,
     "fields": {
       "created_at": "2025-04-02T11:20:00",
@@ -71,7 +95,7 @@
     }
   },
   {
-    "model": "candidates.vacancy",
+    "model": "employers.vacancy",
     "pk": 3,
     "fields": {
       "created_at": "2025-04-03T10:10:00",
@@ -84,7 +108,7 @@
     }
   },
   {
-    "model": "candidates.vacancy",
+    "model": "employers.vacancy",
     "pk": 4,
     "fields": {
       "created_at": "2025-04-04T14:00:00",
@@ -97,7 +121,7 @@
     }
   },
   {
-    "model": "candidates.vacancy",
+    "model": "employers.vacancy",
     "pk": 5,
     "fields": {
       "created_at": "2025-04-05T16:45:00",
@@ -110,7 +134,7 @@
     }
   },
   {
-    "model": "candidates.vacancy",
+    "model": "employers.vacancy",
     "pk": 6,
     "fields": {
       "created_at": "2025-04-06T09:10:00",
@@ -123,7 +147,7 @@
     }
   },
   {
-    "model": "candidates.vacancy",
+    "model": "employers.vacancy",
     "pk": 7,
     "fields": {
       "created_at": "2025-04-06T10:30:00",
@@ -136,7 +160,7 @@
     }
   },
   {
-    "model": "candidates.vacancy",
+    "model": "employers.vacancy",
     "pk": 8,
     "fields": {
       "created_at": "2025-04-06T12:00:00",
@@ -149,7 +173,7 @@
     }
   },
   {
-    "model": "candidates.vacancy",
+    "model": "employers.vacancy",
     "pk": 9,
     "fields": {
       "created_at": "2025-04-06T15:00:00",
@@ -162,7 +186,7 @@
     }
   },
   {
-    "model": "candidates.vacancy",
+    "model": "employers.vacancy",
     "pk": 10,
     "fields": {
       "created_at": "2025-04-06T17:00:00",
