+++ config/settings.py
@@ -37,6 +37,8 @@ INSTALLED_APPS = [
     "django.contrib.messages",
     "django.contrib.staticfiles",
     "candidates",
+    "accounts",
+    "employers",
     "ckeditor",
 ]
