+++ config/urls.py
@@ -16,11 +16,11 @@ Including another URLconf
 """
 from django.conf import settings
 from django.conf.urls.static import static
-# from django.contrib import admin
 from django.urls import path, include
 
 
 urlpatterns = [
     path("", include("candidates.urls", namespace="candidates")),
+    path("employers/", include("employers.urls", namespace="employers")),
     path("accounts/", include("accounts.urls", namespace="accounts")),
 ] + static(settings.STATIC_URL, document_root=settings.STATIC_ROOT)
