+++ templates/candidates/applications.html
@@ -5,7 +5,7 @@
     <div class="container" style="max-width: 630px; font-size: 1.1rem;">
       {% for application in applications %}
         <div class="list-group-item d-flex justify-content-between align-items-center mb-3 py-3 border-top">
-          <a href="{% url 'candidates:jobs_detail' application.vacancy.id %}"
+          <a href="{% url 'employers:vacancies_detail' application.vacancy.id %}"
              class="d-flex w-100 text-dark text-decoration-none">
             <div class="d-flex flex-column">
               <h6 class="mb-1">{{ application.vacancy.title }}</h6>
@@ -14,13 +14,13 @@
           </a>
 
           <div>
-            {% if application.status == 'applied' %}
+            {% if application.status == "applied" %}
               <span class="badge text-muted">Applied</span>
-            {% elif application.status == 'interviewed' %}
+            {% elif application.status == "interviewed" %}
               <span class="badge text-primary">Interviewed</span>
-            {% elif application.status == 'hired' %}
+            {% elif application.status == "hired" %}
               <span class="badge text-success">Hired</span>
-            {% elif application.status == 'rejected' %}
+            {% elif application.status == "rejected" %}
               <span class="badge text-danger">Rejected</span>
             {% endif %}
           </div>
