+++ templates/candidates/index.html
@@ -33,7 +33,7 @@
     </div>
 
     <div class="text-center mb-5">
-      <a href="{% url 'candidates:jobs_list' %}" class="btn btn-dark px-4 py-2 fs-5">Start Exploring Jobs</a>
+      <a href="{% url 'employers:vacancies_list' %}" class="btn btn-dark px-4 py-2 fs-5">Start Exploring Jobs</a>
     </div>
   </div>
 {% endblock %}
