+++ templates/employers/vacancies_detail.html
