+++ templates/employers/vacancies_list.html
@@ -8,7 +8,7 @@
         <div class="col-lg-8">
           {% if vacancies %}
             {% for vacancy in vacancies %}
-              <a href="{% url 'candidates:jobs_detail' vacancy.pk %}"
+              <a href="{% url 'employers:vacancies_detail' vacancy.pk %}"
                  class="vacancy-card d-block mb-4 text-decoration-none text-dark">
                 <div class="bg-white shadow-sm p-5 h-100" style="min-height: 280px; border-radius: 24px;">
                   {% if vacancy.title %}
