+++ templates/includes/filter.html
@@ -41,7 +41,7 @@
                 Apply Filters
               </button>
 
-              <a href="{% url 'candidates:jobs_list' %}" class="flex-fill text-center"
+              <a href="{% url 'employers:vacancies_list' %}" class="flex-fill text-center"
                  style="text-decoration: none; line-height: 40px; border: 1px solid #000; border-radius: 4px;">
                 Reset
               </a>
