+++ templates/includes/footer.html
@@ -3,7 +3,7 @@
     <div class="d-flex justify-content-between align-items-center flex-column flex-md-row position-relative text-center">
 
       <div class="d-flex flex-column align-items-center align-items-md-start gap-2">
-        <a href="{% url 'candidates:jobs_list' %}" class="text-black text-decoration-none">Vacancies</a>
+        <a href="{% url 'employers:vacancies_list' %}" class="text-black text-decoration-none">Vacancies</a>
         <a href="{% url 'candidates:about' %}" class="text-black text-decoration-none">About Us</a>
       </div>
