+++ templates/includes/header.html
@@ -4,7 +4,7 @@
 
       <div class="d-flex align-items-center gap-4">
         <nav class="d-none d-md-flex gap-4">
-          <a href="{% url 'candidates:jobs_list' %}" class="text-black text-decoration-none">Vacancies</a>
+          <a href="{% url 'employers:vacancies_list' %}" class="text-black text-decoration-none">Vacancies</a>
           <a href="{% url 'candidates:about' %}" class="text-black text-decoration-none">About Us</a>
         </nav>
       </div>
@@ -30,7 +30,7 @@
   <hr>
 
   <div class="container d-flex justify-content-center mt-5 px-4" style="max-width: 800px;">
-    <form action="{% url 'candidates:jobs_list' %}" method="get" class="input-group input-group-lg w-100">
+    <form action="{% url 'employers:vacancies_list' %}" method="get" class="input-group input-group-lg w-100">
       <input type="text" name="text" class="form-control border border-dark" placeholder="Search for jobs..."
              style="border-top-left-radius: 999px; border-bottom-left-radius: 999px;" value="{{ search_query }}">
       <button class="btn btn-dark px-4" type="submit"
